+++ .gitignore
@@ -24,10 +24,14 @@
 *.iml
 *.iws
 out/
+classes/
 
 # Maven
 log/
 target/
+
+# .iml files
+.iml
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
