+++ src/main/java/app/dao/CompanyDao.java
@@ -1,6 +1,8 @@
 package app.dao;
 
 import app.entities.Company;
+import app.entities.Employee;
+import app.entities.Project;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -17,5 +19,11 @@ public interface CompanyDao {
 
     void delete(Company company);
 
+    void delete(int id);
+
     void edit(Company company);
+
+    void addEmployeeToCompany(Company company, Employee employee);
+
+    void addProjectToCompany(Company company, Project project);
 }
