+++ src/main/java/app/dao/EmployeeDao.java
@@ -1,6 +1,7 @@
 package app.dao;
 
 import app.entities.Employee;
+import app.entities.Project;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -16,5 +17,9 @@ public interface EmployeeDao {
 
     void delete(Employee employee);
 
+    void delete(int id);
+
     void edit(Employee employee);
+
+    void assignToProject(Employee employee, Project project);
 }
