--- src/main/java/app/resources/ProjectResource.java
@@ -1,55 +1,0 @@
-package app.resources;
-
-import app.entities.Project;
-import app.service.ProjectService;
-import org.springframework.beans.factory.annotation.Autowired;
-
-import javax.ws.rs.*;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import java.util.List;
-
-@Path("/project")
-public class ProjectResource {
-
-    @Autowired
-    private ProjectService projectService;
-
-    @GET
-    @Path("/all")
-    @Produces(MediaType.APPLICATION_JSON)
-    public List<Project> getAll() {
-        return projectService.getProjects();
-    }
-
-    @GET
-    @Path("/{id}")
-    @Produces(MediaType.APPLICATION_JSON)
-    public Project get(@PathParam("id") int id) {
-        return projectService.findById(id);
-    }
-
-    @POST
-    @Path("/add")
-    @Consumes(MediaType.APPLICATION_JSON)
-    public Response add(Project project) {
-        projectService.save(project);
-        return Response.status(Response.Status.CREATED.getStatusCode()).build();
-    }
-
-    @PUT
-    @Path("/edit/{id}")
-    @Consumes(MediaType.APPLICATION_JSON)
-    public Response edit(@PathParam("id") int id, Project project) {
-        projectService.edit(project);
-        return Response.status(Response.Status.CREATED.getStatusCode()).build();
-    }
-
-    @DELETE
-    @Path("/delete/{id}")
-    @Consumes(MediaType.APPLICATION_JSON)
-    public Response delete(@PathParam("id") int id) {
-        projectService.delete(id);
-        return Response.status(Response.Status.OK.getStatusCode()).build();
-    }
-}
