+++ src/main/java/app/service/EmployeeService.java
@@ -1,25 +1,68 @@
 package app.service;
 
+import app.dao.EmployeeDao;
+import app.dao.ProjectDao;
 import app.entities.Employee;
-import app.entities.Project;
-import org.springframework.stereotype.Service;
+import org.springframework.beans.factory.annotation.Autowired;
 
+import javax.ws.rs.*;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 import java.util.List;
 
-@Service
-public interface EmployeeService {
+@Path("/employee")
+public class EmployeeService {
 
-    void assignToProject(Employee employee, Project project);
+    @Autowired
+    private EmployeeDao employeeDao;
+    @Autowired
+    private ProjectDao projectDao;
 
-    List<Employee> getEmployees();
+    @GET
+    @Path("/all")
+    @Produces(MediaType.APPLICATION_JSON)
+    public List<Employee> getAll() {
+        return employeeDao.getAll();
+    }
 
-    Employee findById(int id);
+    @GET
+    @Path("/{id}")
+    @Produces(MediaType.APPLICATION_JSON)
+    public Employee get(@PathParam("id") int id) {
+        return employeeDao.findById(id);
+    }
 
-    void save(Employee employee);
+    @POST
+    @Path("/add")
+    @Consumes(MediaType.APPLICATION_JSON)
+    public Response add(Employee employee) {
+        employeeDao.save(employee);
+        return Response.status(Response.Status.CREATED.getStatusCode()).build();
+    }
 
-    void delete(Employee employee);
+    @PUT
+    @Path("/edit/{id}")
+    @Consumes(MediaType.APPLICATION_JSON)
+    public Response edit(@PathParam("id") int id, Employee employee) {
+        employeeDao.edit(employee);
+        return Response.status(Response.Status.CREATED.getStatusCode()).build();
+    }
 
-    void delete(int id);
+    @DELETE
+    @Path("/delete/{id}")
+    @Consumes(MediaType.APPLICATION_JSON)
+    public Response delete(@PathParam("id") int id) {
+        employeeDao.delete(id);
+        return Response.status(Response.Status.OK.getStatusCode()).build();
+    }
 
-    void edit(Employee employee);
+    @POST
+    @Path("/assign/{employeeId}/{projectId}")
+    @Consumes(MediaType.APPLICATION_JSON)
+    public Response assignToProject(@PathParam("employeeId") int employeeId,
+                                    @PathParam("projectId") int projectId) {
+        employeeDao.assignToProject(employeeDao.findById(employeeId),
+                projectDao.findById(projectId));
+        return Response.status(Response.Status.OK.getStatusCode()).build();
+    }
 }
