+++ src/main/java/app/service/LogsService.java
@@ -1,19 +1,63 @@
 package app.service;
 
+
+import app.dao.LogsDao;
 import app.entities.Logs;
 import app.entities.namespace.LogsNamespace;
-import org.springframework.stereotype.Service;
+import org.springframework.beans.factory.annotation.Autowired;
 
+import javax.ws.rs.*;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 import java.util.List;
 
-@Service
-public interface LogsService {
+@Path("/logs")
+public class LogsService {
 
-    List<Logs> getLogFor(LogsNamespace logsNamespace);
+    @Autowired
+    private LogsDao logsDao;
 
-    List<Logs> getAllLogs();
+    @GET
+    @Path("/all")
+    @Produces(MediaType.APPLICATION_JSON)
+    public List<Logs> getAll() {
+        return logsDao.getAll();
+    }
 
-    void save(Logs logs);
+    @GET
+    @Path("/get/today")
+    @Produces(MediaType.APPLICATION_JSON)
+    public List<Logs> getForToday() {
+        return logsDao.getLogFor(LogsNamespace.TODAY);
+    }
 
-    void delete(Logs logs);
+    @GET
+    @Path("/get/week")
+    @Produces(MediaType.APPLICATION_JSON)
+    public List<Logs> getForThisWeek() {
+        return logsDao.getLogFor(LogsNamespace.THIS_WEEK);
+    }
+
+    @GET
+    @Path("/get/month")
+    @Produces(MediaType.APPLICATION_JSON)
+    public List<Logs> getForThisMonth() {
+        return logsDao.getLogFor(LogsNamespace.THIS_MONTH);
+    }
+
+    @POST
+    @Path("/add")
+    @Consumes(MediaType.APPLICATION_JSON)
+    public Response add(Logs logs) {
+        logsDao.save(logs);
+        return Response.status(Response.Status.CREATED.getStatusCode()).build();
+    }
+
+    @DELETE
+    @Path("/delete")
+    @Consumes(MediaType.APPLICATION_JSON)
+    public Response delete(Logs logs) {
+        logsDao.delete(logs);
+        return Response.status(Response.Status.OK.getStatusCode()).build();
+    }
 }
